fix warning for type comparison

+++ src/navigation_test.h
@@ -50,8 +50,8 @@ inline int Simulate(int argc, const char** argv) {
   // print navigation_map
   bool print_navigation_map = false;
   if (print_navigation_map) {
-    for (int map_x = 0; map_x < navigation_map.size(); map_x++) {
-      for (int map_y = 0; map_y < navigation_map[0].size(); map_y++) {
+    for (size_t map_x = 0; map_x < navigation_map.size(); map_x++) {
+      for (size_t map_y = 0; map_y < navigation_map[0].size(); map_y++) {
         std::cout << navigation_map[map_x][map_y];
       }
       std::cout << std::endl;
